fix suggestion
move useEffect to get suggestion to component AddTag
